update: convert `UserImage` component to reusable one. implement it into `DashboardHeader`
